Oppdatert objekter for særbidrag (sjabloner + barn i husstand) (#171)

+++ bidrag-transport/src/main/kotlin/no/nav/bidrag/transport/behandling/felles/grunnlag/Beregning.kt
@@ -36,7 +36,7 @@ data class DelberegningSumInntekt(
 
 data class DelberegningBarnIHusstand(
     override val periode: ÅrMånedsperiode,
-    val antallBarn: Int,
+    val antallBarn: Double,
 ) : Delberegning
 
 data class DelberegningBidragsevne(
